+++ src/azure-cli/azure/cli/command_modules/cdn/aaz/latest/cdn/custom_domain/_delete.py
