+++ src/azure-cli/azure/cli/command_modules/cdn/tests/latest/scenario_mixin.py
